Fix some of the many issues in this spec
Note leaving the submit_and_verify_alert method untouched for now as it
was broken before and I do not have time to fix it as part of this
sprint. It should be fix soon though. I have marked it pending for now
because it will not pass until that is fixed

[INT-35]