Wait for login to occur before continuing
Sometimes, after filling out the Perry login form, it can take a
few seconds for the authorization to happen and the page to
successfully redirect. If we immediately try to navigate to a page
within the app, we may not actually have gotten a token, causing
our login to effectively fail. This change adds a delay, waiting
until we are on some page that does not have a Sign In button (in
this case the dashboard).